update react query to tanstack react query, move providers to index
