+++ src/frontend/OverlayManager/index.tsx
@@ -8,11 +8,7 @@ import { BrowserGameProps } from './types'
 import { Overlay } from './Overlay'
 import { WebviewTag } from 'electron'
 import WebviewControls from 'frontend/components/UI/WebviewControls'
-import { WagmiProvider } from 'wagmi'
-import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
-import { config } from './config'
 import AuthModal from 'frontend/components/UI/AuthModal'
-const queryClient = new QueryClient()
 
 const OverlayManager = observer(function ({
   appName,
@@ -40,44 +36,42 @@ const OverlayManager = observer(function ({
 
   /* eslint-disable react/no-unknown-property */
   return (
-    <WagmiProvider config={config}>
-      <QueryClientProvider client={queryClient}>
-        <AuthModal />
-        <div
-          className={BrowserGameStyles.overlayContainer}
-          style={style}
-          id="overlay-manager"
-        >
-          {OverlayState.showOverlay ? (
-            <Overlay appName={appName} runner={runner} />
-          ) : null}
-          {url !== 'ignore' && OverlayState.renderState.showBrowserGame ? (
-            <div>
-              <WebviewControls
-                webview={webviewRef.current}
-                initURL={''}
-                openInBrowser={false}
-                disableUrl={true}
-              />
-              <webview
-                src={url}
-                className={BrowserGameStyles.browserGame}
-                partition={
-                  WalletState.provider === PROVIDERS.METAMASK_MOBILE ||
-                  PROVIDERS.WALLET_CONNECT
-                    ? 'persist:InPageWindowEthereumExternalWallet'
-                    : undefined
-                }
-                webpreferences="contextIsolation=true"
-                // setting = to {true} does not work :(
-                allowpopups={trueAsStr}
-                ref={webviewRef}
-              />
-            </div>
-          ) : null}
-        </div>
-      </QueryClientProvider>
-    </WagmiProvider>
+    <>
+      <AuthModal />
+      <div
+        className={BrowserGameStyles.overlayContainer}
+        style={style}
+        id="overlay-manager"
+      >
+        {OverlayState.showOverlay ? (
+          <Overlay appName={appName} runner={runner} />
+        ) : null}
+        {url !== 'ignore' && OverlayState.renderState.showBrowserGame ? (
+          <div>
+            <WebviewControls
+              webview={webviewRef.current}
+              initURL={''}
+              openInBrowser={false}
+              disableUrl={true}
+            />
+            <webview
+              src={url}
+              className={BrowserGameStyles.browserGame}
+              partition={
+                WalletState.provider === PROVIDERS.METAMASK_MOBILE ||
+                PROVIDERS.WALLET_CONNECT
+                  ? 'persist:InPageWindowEthereumExternalWallet'
+                  : undefined
+              }
+              webpreferences="contextIsolation=true"
+              // setting = to {true} does not work :(
+              allowpopups={trueAsStr}
+              ref={webviewRef}
+            />
+          </div>
+        ) : null}
+      </div>
+    </>
   )
 })
