+++ src/frontend/hooks/useGetSteamGame.ts
@@ -1,12 +1,12 @@
-import { useQueries } from 'react-query'
+import { useQueries } from '@tanstack/react-query'
 
 export default function useGetSteamGame(
   steam_games: {
     id: string
   }[]
 ) {
-  const query = useQueries(
-    steam_games.map((val) => ({
+  const query = useQueries({
+    queries: steam_games.map((val) => ({
       queryKey: ['getSteamGame', val.id],
       queryFn: async () => {
         const response = await window.api.getSteamGameMetadata(
@@ -20,7 +20,7 @@ export default function useGetSteamGame(
       },
       staleTime: Infinity
     }))
-  )
+  })
 
   return {
     data: query
